Se cambiaron de lugar algunas clases y se comenzó la refactorización de la clase Animats.
