+++ .gitignore
@@ -1,6 +1,7 @@
 # Build and Release Folders
 bin-debug/
 bin-release/
+bin/
 
 # Project property files
 .actionScriptProperties
